# swagger.yaml
openapi: 3.0.2
info:
  title: FastAPI
  version: 0.1.0
paths:
...
  /auth/refresh:
    post:
      tags:
        - auth
      summary: Refresh token
      operationId: refresh_token
      requestBody:
        content:
          application/json:
            schema:
              properties:
                grant_type:
                  title: The status of the operation
                  type: string
                  default: refresh_token
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
...